requirements.txt
datasets==4.1.1
faiss==1.9.0
invoke==2.2.0
numpy==2.3.3
pandas==2.3.2
pytest==8.4.2
PyYAML==6.0.3
sentence_transformers==5.1.1
torch==2.6.0
tqdm==4.67.1
wandb
loguru
python-box
